Rename module pyk.subst to pyk.rewrite
